Improves crash dump
On crash we're somtimes stuck in the alternate screen in raw mode. The
error prints and the screen then resets when the printing is done. This
prevents the user from reading the error message.

We now attempt to switch to the main buffer and reset before printing
the error message. Also the message is printed in a RAW mode friendly
way.