Improves "TTS not available" handling and feedback
Until now, when blightmud is compiled without TTS support the feedback
that TTS isn't available doesn't really exist. With this change.
Attempting to enable TTS via commandline or through lua scripts
will present the user with proper warnings that his/her version of
Blightmud is not compiled with TTS enabled.