deps: update ndk-sys 0.4.0 -> 0.4.1.
ndk-sys 0.4.0 is a transitive dependency of Blightmud. Unfortunately
a known issue with the cargo registry and how it applies semver has left
the `ndk-sys` crate producing checksum mismatche errors between a 0.4.0
release and a yanked 0.4.0+25.0.8775105 release. This manifest as the
build error:

  Failed to verify checksum of ndk-sys v0.4.0

The author of the ndk-sys crate has published a 0.4.1 release that
should resolve this build error. This commit updates the `Cargo.lock` to
use this version.