prompt: add add_prompt_listener function.
This commit adds a new `add_prompt_listener` function to the Lua prompt
module. This function allows registering a callback that will be
invoked whenever the input prompt buffer area is updated (either by the
user typing, or from `prompt.set()` being invoked).

In the future this may be useful for performing tasks like
spellchecking of data entered in the prompt area but not yet sent to the
server.