feat: spellchecking API.
This commit adds a new Lua module, `spellcheck`, with three functions:

* `spellcheck.init(aff_path, dict_path)` - initializes a spellchecker
  with the paths to a Hunspell[0] compatible affix file and dictionary
  file. Must be called before any other spellcheck functions or an error
  will be thrown.
* `spellcheck.check(word) -> bool` - returns true if the word is in the
  dictionary that the spellchecker was initialized with. Errs if the
  spellchecker hasn't been initialized.
* `spellcheck.suggest(word) -> table` - if the given word is misspelled,
  returns a table array of suggested replacements in order of
  likelihood.

Presently these features are implemented with a backend that relies on
the `hunspell-rs` crate[1], which in turn uses `hunspell-sys`[2] to
interact with libhunspell[0]. Hunspell is used extensively (e.g. by
LibreOffice, Firefox) so compatible dictionaries/affix files are easy
to source and the results are competitive with popular software. To
avoid adding a new external dependency we use the "bundled" feature of
`hunspell-sys` to have that crate bring along its own statically linked
libhunspell. This means no changes are required for CI or the release
process.

In the future the `zspell`[3] crate looks like a promising pure Rust
alternative to `hunspell-rs` that supports the Hunspell format and API.
I elected not to use that crate for the short term because it doesn't
yet have a stable API for providing suggestions and that's a core
feature for a spellchecker. It should be possible to switch to this
crate transparent to end users once it's feature complete.

For an example interactive spellchecker plugin built with this new Lua
module, see Blightspell[4].

[0]: http://hunspell.github.io/
[1]: https://crates.io/crates/hunspell-rs
[2]: https://crates.io/crates/hunspell-sys
[3]: https://crates.io/crates/zspell
[4]: https://github.com/cpu/blightspell