Prevent Event::SetPromptMask from interupting ctrl-c
The `ctrl-c` cycle (needing to submit it twice) is sensitive to other
events arriving between the first and second `ctrl-c` (sigint) signal.
The new event SetPromptMask was not included in the whitelist of events
that can be received and would interupt the ctrl-c sequence.

This system should probably be re-worked to use some form of timer or at
least highlight that the sequence has been interrupted. It's otherwise
quite easy for a plugin to unknowingly interrupt this cycle.