net: rewrite version check using reqwest.
Previously the `check_version.rs` implementation relied on Rust bindings
for the native `curl` library. This was the only place in the codebase
using that dependency.

Instead of relying on native code, this commit replaces the
implementation with `reqwest`[0] using the `rustls` TLS backend we
implemented for MUD TLS. Using reqwest provides end-to-end HTTP(S)
requests in pure Rust.

Since we're performing version checks in a background thread I've
implemented this check using the blocking reqwest client.

Along the way I made some refactoring cleanups:

* `diff_version` isn't required, we can use `&str.cmp` directly.
* `Fetcher`'s error `Result` was never meaningfully used, so switch to
  `Option`.
* Rather than returning `Vec<u8>` and manually deserializing fields, the
  `Fetcher` trait is switched to return a struct with a derived `Deserialize`
  and the impl uses reqwest's `Response.json` to unmarshal the HTTP
  response into it.

[0]: https://docs.rs/reqwest/latest/reqwest/