lua: tidy up lua resource loading, fix chunk name.
Previously the `create_default_lua_state` fn from `lua_script.rs` had
a fair bit of duplication loading built-in lua resources. The existing
code also did not set the chunk name with the path of the resource being
loaded, meaning if there was an error in the resource script the
location would be described within `lua_script.rs` rather than within
the resource script.

Similarly, the handler for `/load`ing a Lua script was setting the chunk
name to the parent directory of the loaded script (or "" if there wasn't
one) instead of the path to the script itself. This leads to confusing
errors if loading a script from the PWD when there is no parent dir.

To resolve the first issue this commit adds two new `macro_rules!`:

* `lua_global_resources` for loading one or more resource scripts that
  should be called and stored in the globals.
* `lua_resources` for loading one or more resource scripts and exec'ing
  it.

Both macros take care to set the chunk name to the resource script path
for better error messages.

To resolve the second issue the load script handler is fixed to set the
chunk name to the script path instead of the parent directory.