robust ctrl-c quit confirmation.
Previously the quit confirmation that made sure users didn't exit
Blightmud unexpectedly was based on looking at events received between
the first ctrl-c press and the confirmation. This required keeping an
allow list of ignored events (UserInputBuffer, TimedEvent, TimerTick,
SetPromptMask, and ClearPromptMask) that shouldn't clear the
confirmation flag when received. In practice keeping that list up to
date as new event types are added is problematic and so a more robust
alternative is needed.

This commit replaces the quit confirmation with a new system where the
confirmation is cleared by a dedicated event (QuitConfirmTimeout)
sent by a separate thread that sleeps for 5s. The first ctrl-c sets the
pending quit flag. A second ctrl-c being received before the timeout
thread wakes up will quit Blightmud. If the thread wakes up before
a second confirmation ctrl-c is sent then it writes a QuitConfirmTimeout
event that will result in the pending quit flag being cleared. If the
user presses ctrl-c at this point it will start the cycle again.