feature: make spellcheck dependency optional.
This commit adds a `spellcheck` feature flag that is default-enabled,
and makes the `hunspell-rs` and `hunspell-sys` dependencies optional
based on that feature flag.

This will allow most users to get spellchecking by default, but for the
select few having trouble building `hunspell-sys` on ARM64, they can use
`--no-default-features` to avoid the problem.