feat: add echo_input setting.
Previously all `ServerInput` events unconditionally called
`screen.print_send` to echo input sent by a user to the screen. In the
split screen mode this results in the input appearing with a `> ` prefix
and special colouring.

This commit adds a `echo_input` setting that can be changed to disable
screen input echo behaviour on an opt-in basis. By default `echo_input`
is true, matching existing behaviour.

It can be changed at runtime with immediate effect. When set to false
input is still logged but will no longer appear on the screen with a `>
` prefix. Setting it to true will restore the echo effect.

Since this feature disables echoing of input it is mutually exclusive
with the ability to use `search.find_last_input`,
`search.find_next_input` and the `Ctrl PgUp`/`Ctrl PgDown` bindings. For
that reason the respective resources files are updated to note that this
is only possible with `/set echo_input on`. The `search` lua module is
updated to output an error message if `find_(next|last)_input` are
called when `echo_input` is off.