Cargo: specify opt-level for hunspell-sys
The `hunspell-sys` crate builds native code that emits warnings when
building with `--debug`. These warnings are treated as errors by cargo
and so breaks blightmud builds that enable spellcheck unless `--release`
is used.

This commit fixes `--debug` builds by explicitly setting optimizations
enabled for the `hunspell-sys` dependency when using the `dev` profile.
We don't need debug info for hunspell-sys and this fixes the build
error.