lua: support emoji in GMCP data
This commit updates the `gmcp.lua`'s `_subneg_recv` function so that
when given a `data` table containing the byte values of a UTF-8 string
that contains multi-byte characters (e.g. emoji), the values are
correctly preserved and passed on to the GMCP handler.

Using `utf8.char(unpack(data))` was corrupting the emoji, resulting in
misc garbled symbols in its place.