net: configure TCP keepalive, more debug info
This commit updates the `open_tcp_stream` fn from `net/util.rs` to
configure the opened socket with TCP keepalive[0] enabled.

This TCP feature is particularly important for MUDs because Telnet has
no consistently enabled application layer keepalive feature. Without
a keepalive mechanism long-lived and idle Telnet connections are prone
to being closed by middlebox routers. When this happens Blightmud won't
detect the socket was closed until the user tries to read or write from
the socket. Other MUD clients like Mudlet configure TCP keepalive to
compensate for this using a native (but typically default off) TCP
feature.

In order to achieve TCP keepalive for Blightmud we take a new dependency
on `socket2`[1], a crate that abstracts over unsafe OS-specific
mechanisms for customizing socket options. We use `socket2` to configure
the opened socket to start sending TCP keepalives 30 seconds after
a connection is detected as idle. Keepalives will be sent every
5 seconds, and we'll tolerate up to 5 lost keepalives before deciding
the connection is closed.

We default to configuring TCP keepalive but a new "keepalive_enabled"
setting can be set to 'false' by users that don't want this feature for
some reason.

Lastly, to aid in future debugging new `debug!` log lines are added for
which addresses are resolved, which are tried, and errors encountered
trying to connect. Using the debug log level means these won't be
present unless Blightmud is run with `-V` for verbose mode.

[0]: https://en.wikipedia.org/wiki/Keepalive#TCP_keepalive
[1]: https://docs.rs/socket2/latest/socket2/