net: rework open_tcp_stream keepalive test
Previously the `test_keepalive_disable` unit test used `Settings` to
realistically enable/disable the keepalive setting and then testing that
the created TCP stream/socket were configured correctly based on the
setting state. This is nice in the sense that it matches real world
usage, but the filesystem is a shared resource and unit tests are run in
parallel by default.

This means this test had bad interactions with other tests
reading/writing settings from disk (particularly the v2 migration tests)
since they all use the same data dir and config dir (under `.run/` in
the crate directory). This creates test flakes and makes maintainers sad
:(

Since making the `Settings` class easier to use in parallel from
multiple tests will be a larger project this commit refactors the
existing test to not rely on the `Settings` interface. This will remove
the flakyness of the tests and we can consider broader changes
separately.