net: configure Rustls w/ KeyLogFile support
This commit updates the Rustls client configuration Blightmud uses for
TLS sessions to write a key log file to disk when a special
"SSLKEYLOGFILE" env var is set to a file path.

This is important to enable sharing pcaps of TLS data with Blightmud
maintainers to help troubleshoot bugs. Without the TLS key log file the
pcap is almost entirely opaque encrypted data.