net: log mud receiver stream read errors
Previously the `MudReceiver` logic that could potentially read an
`Error` result from `self.connection.read(&mut data)` only used the `Ok`
result. If there was an error, `data` is set to the empty vec and
a higher layer treats the empty data as an EOF initiated disconnect.

This commit breaks out the result handling to log a read error with
`error!`, and otherwise maintains the existing behaviour: returning an
empty vec.