fix TLS, remove manual test provider management
Use just ring for the crypto provider and let the unambiguous default be
used by rustls as the process wide provider.

Previously Blightmud ended up in the situation where it was relying on
both the ring cryptography provider, and the aws-lc-rs provider. Because
it was ambiguous which provider should be the default in this
arrangement the unit tests had to explicit set the default provider to
aws-lc-rs. Unfortunately since nothing did the same thing for _non-test_
contexts, Blightmud will panic on the first TLS connection.

We can fix this by making the default unambiguous: settling on a single
crypto provider and avoiding any manual configuration.

The reason two providers were being used is a bit subtle:

1. The dangerous certificate verifier was ref'ing ring explicitly.
2. The reqwest dep changed its default to use aws-lc-rs.

In practice for Blightmud users I think the pre-existing choice (ring)
will be less friction and so I've restored it as the unambiguous default
and removed the manual test-context provider setting logic.

Thus the fix is to make the dangerous verifier use whatever the default
provider is, and to make the default provider unambiguous by setting the
right rustls+reqwest feature flags.